+++ src/main/kotlin/no/nav/familie/ba/sak/kjerne/vedtak/vedtaksperiode/vedtaksperiodeProdusent/Eksempel sammenslåing avslagslagsperioder med andre perioder.png
