+++ src/main/kotlin/no/nav/familie/ba/sak/kjerne/vedtak/vedtaksperiode/vedtaksperiodeProdusent/GrunnlagForGjeldendeOgForrigeBehandling.kt
@@ -1,4 +1,4 @@
-package no.nav.familie.ba.sak.kjerne.vedtak.vedtaksperiode.produsent
+package no.nav.familie.ba.sak.kjerne.vedtak.vedtaksperiode.vedtaksperiodeProdusent
 
 data class GrunnlagForGjeldendeOgForrigeBehandling(
     val gjeldende: VedtaksperiodeGrunnlagForPerson?,
